Cut over remaining queries to registerQuery
