+++ core/src/main/java/org/elasticsearch/index/query/BoostingQueryBuilder.java
@@ -21,9 +21,12 @@ package org.elasticsearch.index.query;
 
 import org.apache.lucene.queries.BoostingQuery;
 import org.apache.lucene.search.Query;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -43,6 +46,12 @@ import java.util.Objects;
 public class BoostingQueryBuilder extends AbstractQueryBuilder<BoostingQueryBuilder> {
 
     public static final String NAME = "boosting";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final BoostingQueryBuilder PROTOTYPE = new BoostingQueryBuilder(EmptyQueryBuilder.PROTOTYPE, EmptyQueryBuilder.PROTOTYPE);
+
+    private static final ParseField POSITIVE_FIELD = new ParseField("positive");
+    private static final ParseField NEGATIVE_FIELD = new ParseField("negative");
+    private static final ParseField NEGATIVE_BOOST_FIELD = new ParseField("negative_boost");
 
     private final QueryBuilder positiveQuery;
 
@@ -50,7 +59,6 @@ public class BoostingQueryBuilder extends AbstractQueryBuilder<BoostingQueryBuil
 
     private float negativeBoost = -1;
 
-    static final BoostingQueryBuilder PROTOTYPE = new BoostingQueryBuilder(EmptyQueryBuilder.PROTOTYPE, EmptyQueryBuilder.PROTOTYPE);
 
     /**
      * Create a new {@link BoostingQueryBuilder}
@@ -104,13 +112,70 @@ public class BoostingQueryBuilder extends AbstractQueryBuilder<BoostingQueryBuil
     @Override
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
-        builder.field(BoostingQueryParser.POSITIVE_FIELD.getPreferredName());
+        builder.field(POSITIVE_FIELD.getPreferredName());
         positiveQuery.toXContent(builder, params);
-        builder.field(BoostingQueryParser.NEGATIVE_FIELD.getPreferredName());
+        builder.field(NEGATIVE_FIELD.getPreferredName());
         negativeQuery.toXContent(builder, params);
-        builder.field(BoostingQueryParser.NEGATIVE_BOOST_FIELD.getPreferredName(), negativeBoost);
+        builder.field(NEGATIVE_BOOST_FIELD.getPreferredName(), negativeBoost);
         printBoostAndQueryName(builder);
         builder.endObject();
+    }
+
+    public static BoostingQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+
+        QueryBuilder positiveQuery = null;
+        boolean positiveQueryFound = false;
+        QueryBuilder negativeQuery = null;
+        boolean negativeQueryFound = false;
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+        float negativeBoost = -1;
+        String queryName = null;
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, POSITIVE_FIELD)) {
+                    positiveQuery = parseContext.parseInnerQueryBuilder();
+                    positiveQueryFound = true;
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, NEGATIVE_FIELD)) {
+                    negativeQuery = parseContext.parseInnerQueryBuilder();
+                    negativeQueryFound = true;
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[boosting] query does not support [" + currentFieldName + "]");
+                }
+            } else if (token.isValue()) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, NEGATIVE_BOOST_FIELD)) {
+                    negativeBoost = parser.floatValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                    queryName = parser.text();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                    boost = parser.floatValue();
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[boosting] query does not support [" + currentFieldName + "]");
+                }
+            }
+        }
+
+        if (!positiveQueryFound) {
+            throw new ParsingException(parser.getTokenLocation(), "[boosting] query requires 'positive' query to be set'");
+        }
+        if (!negativeQueryFound) {
+            throw new ParsingException(parser.getTokenLocation(), "[boosting] query requires 'negative' query to be set'");
+        }
+        if (negativeBoost < 0) {
+            throw new ParsingException(parser.getTokenLocation(),
+                    "[boosting] query requires 'negative_boost' to be set to be a positive value'");
+        }
+
+        BoostingQueryBuilder boostingQuery = new BoostingQueryBuilder(positiveQuery, negativeQuery);
+        boostingQuery.negativeBoost(negativeBoost);
+        boostingQuery.boost(boost);
+        boostingQuery.queryName(queryName);
+        return boostingQuery;
     }
 
     @Override
