+++ core/src/main/java/org/elasticsearch/index/query/SpanContainingQueryBuilder.java
@@ -22,9 +22,12 @@ package org.elasticsearch.index.query;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.spans.SpanContainingQuery;
 import org.apache.lucene.search.spans.SpanQuery;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -36,10 +39,15 @@ public class SpanContainingQueryBuilder extends AbstractQueryBuilder<SpanContain
         implements SpanQueryBuilder<SpanContainingQueryBuilder> {
 
     public static final String NAME = "span_containing";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final SpanContainingQueryBuilder PROTOTYPE =
+            new SpanContainingQueryBuilder(SpanTermQueryBuilder.PROTOTYPE, SpanTermQueryBuilder.PROTOTYPE);
+
+    private static final ParseField BIG_FIELD = new ParseField("big");
+    private static final ParseField LITTLE_FIELD = new ParseField("little");
+
     private final SpanQueryBuilder big;
     private final SpanQueryBuilder little;
-    static final SpanContainingQueryBuilder PROTOTYPE =
-            new SpanContainingQueryBuilder(SpanTermQueryBuilder.PROTOTYPE, SpanTermQueryBuilder.PROTOTYPE);
 
     /**
      * @param big the big clause, it must enclose {@code little} for a match.
@@ -73,12 +81,56 @@ public class SpanContainingQueryBuilder extends AbstractQueryBuilder<SpanContain
     @Override
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
-        builder.field(SpanContainingQueryParser.BIG_FIELD.getPreferredName());
+        builder.field(BIG_FIELD.getPreferredName());
         big.toXContent(builder, params);
-        builder.field(SpanContainingQueryParser.LITTLE_FIELD.getPreferredName());
+        builder.field(LITTLE_FIELD.getPreferredName());
         little.toXContent(builder, params);
         printBoostAndQueryName(builder);
         builder.endObject();
+    }
+
+    public static SpanContainingQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+        String queryName = null;
+        SpanQueryBuilder<?> big = null;
+        SpanQueryBuilder<?> little = null;
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, BIG_FIELD)) {
+                    QueryBuilder query = parseContext.parseInnerQueryBuilder();
+                    if (!(query instanceof SpanQueryBuilder<?>)) {
+                        throw new ParsingException(parser.getTokenLocation(), "span_containing [big] must be of type span query");
+                    }
+                    big = (SpanQueryBuilder<?>) query;
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, LITTLE_FIELD)) {
+                    QueryBuilder query = parseContext.parseInnerQueryBuilder();
+                    if (!(query instanceof SpanQueryBuilder<?>)) {
+                        throw new ParsingException(parser.getTokenLocation(), "span_containing [little] must be of type span query");
+                    }
+                    little = (SpanQueryBuilder<?>) query;
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(),
+                            "[span_containing] query does not support [" + currentFieldName + "]");
+                }
+            } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                boost = parser.floatValue();
+            } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                queryName = parser.text();
+            } else {
+                throw new ParsingException(parser.getTokenLocation(),
+                        "[span_containing] query does not support [" + currentFieldName + "]");
+            }
+        }
+
+        SpanContainingQueryBuilder query = new SpanContainingQueryBuilder(big, little);
+        query.boost(boost).queryName(queryName);
+        return query;
     }
 
     @Override
