+++ core/src/main/java/org/elasticsearch/index/query/SpanNearQueryBuilder.java
@@ -22,9 +22,12 @@ package org.elasticsearch.index.query;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.spans.SpanNearQuery;
 import org.apache.lucene.search.spans.SpanQuery;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -39,17 +42,22 @@ import java.util.Objects;
 public class SpanNearQueryBuilder extends AbstractQueryBuilder<SpanNearQueryBuilder> implements SpanQueryBuilder<SpanNearQueryBuilder> {
 
     public static final String NAME = "span_near";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final SpanNearQueryBuilder PROTOTYPE = new SpanNearQueryBuilder(SpanTermQueryBuilder.PROTOTYPE, 0);
 
     /** Default for flag controlling whether matches are required to be in-order */
     public static boolean DEFAULT_IN_ORDER = true;
+
+    private static final ParseField SLOP_FIELD = new ParseField("slop");
+    private static final ParseField COLLECT_PAYLOADS_FIELD = new ParseField("collect_payloads").withAllDeprecated("no longer supported");
+    private static final ParseField CLAUSES_FIELD = new ParseField("clauses");
+    private static final ParseField IN_ORDER_FIELD = new ParseField("in_order");
 
     private final List<SpanQueryBuilder<?>> clauses = new ArrayList<>();
 
     private final int slop;
 
     private boolean inOrder = DEFAULT_IN_ORDER;
-
-    static final SpanNearQueryBuilder PROTOTYPE = new SpanNearQueryBuilder(SpanTermQueryBuilder.PROTOTYPE, 0);
 
     /**
      * @param initialClause an initial span query clause
@@ -105,15 +113,79 @@ public class SpanNearQueryBuilder extends AbstractQueryBuilder<SpanNearQueryBuil
     @Override
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
-        builder.startArray(SpanNearQueryParser.CLAUSES_FIELD.getPreferredName());
+        builder.startArray(CLAUSES_FIELD.getPreferredName());
         for (SpanQueryBuilder<?> clause : clauses) {
             clause.toXContent(builder, params);
         }
         builder.endArray();
-        builder.field(SpanNearQueryParser.SLOP_FIELD.getPreferredName(), slop);
-        builder.field(SpanNearQueryParser.IN_ORDER_FIELD.getPreferredName(), inOrder);
+        builder.field(SLOP_FIELD.getPreferredName(), slop);
+        builder.field(IN_ORDER_FIELD.getPreferredName(), inOrder);
         printBoostAndQueryName(builder);
         builder.endObject();
+    }
+
+    public static SpanNearQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+        Integer slop = null;
+        boolean inOrder = SpanNearQueryBuilder.DEFAULT_IN_ORDER;
+        String queryName = null;
+
+        List<SpanQueryBuilder> clauses = new ArrayList<>();
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_ARRAY) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, CLAUSES_FIELD)) {
+                    while ((token = parser.nextToken()) != XContentParser.Token.END_ARRAY) {
+                        QueryBuilder query = parseContext.parseInnerQueryBuilder();
+                        if (!(query instanceof SpanQueryBuilder)) {
+                            throw new ParsingException(parser.getTokenLocation(), "spanNear [clauses] must be of type span query");
+                        }
+                        clauses.add((SpanQueryBuilder) query);
+                    }
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[span_near] query does not support [" + currentFieldName + "]");
+                }
+            } else if (token.isValue()) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, IN_ORDER_FIELD)) {
+                    inOrder = parser.booleanValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, COLLECT_PAYLOADS_FIELD)) {
+                    // Deprecated in 3.0.0
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, SLOP_FIELD)) {
+                    slop = parser.intValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                    boost = parser.floatValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                    queryName = parser.text();
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[span_near] query does not support [" + currentFieldName + "]");
+                }
+            } else {
+                throw new ParsingException(parser.getTokenLocation(), "[span_near] query does not support [" + currentFieldName + "]");
+            }
+        }
+
+        if (clauses.isEmpty()) {
+            throw new ParsingException(parser.getTokenLocation(), "span_near must include [clauses]");
+        }
+
+        if (slop == null) {
+            throw new ParsingException(parser.getTokenLocation(), "span_near must include [slop]");
+        }
+
+        SpanNearQueryBuilder queryBuilder = new SpanNearQueryBuilder(clauses.get(0), slop);
+        for (int i = 1; i < clauses.size(); i++) {
+            queryBuilder.clause(clauses.get(i));
+        }
+        queryBuilder.inOrder(inOrder);
+        queryBuilder.boost(boost);
+        queryBuilder.queryName(queryName);
+        return queryBuilder;
     }
 
     @Override
