+++ core/src/main/java/org/elasticsearch/index/query/SpanNotQueryBuilder.java
@@ -22,9 +22,12 @@ package org.elasticsearch.index.query;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.spans.SpanNotQuery;
 import org.apache.lucene.search.spans.SpanQuery;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -32,11 +35,20 @@ import java.util.Objects;
 public class SpanNotQueryBuilder extends AbstractQueryBuilder<SpanNotQueryBuilder> implements SpanQueryBuilder<SpanNotQueryBuilder> {
 
     public static final String NAME = "span_not";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final SpanNotQueryBuilder PROTOTYPE = new SpanNotQueryBuilder(SpanTermQueryBuilder.PROTOTYPE,
+            SpanTermQueryBuilder.PROTOTYPE);
 
     /** the default pre parameter size */
     public static final int DEFAULT_PRE = 0;
     /** the default post parameter size */
     public static final int DEFAULT_POST = 0;
+
+    private static final ParseField POST_FIELD = new ParseField("post");
+    private static final ParseField PRE_FIELD = new ParseField("pre");
+    private static final ParseField DIST_FIELD = new ParseField("dist");
+    private static final ParseField EXCLUDE_FIELD = new ParseField("exclude");
+    private static final ParseField INCLUDE_FIELD = new ParseField("include");
 
     private final SpanQueryBuilder include;
 
@@ -45,8 +57,6 @@ public class SpanNotQueryBuilder extends AbstractQueryBuilder<SpanNotQueryBuilde
     private int pre = DEFAULT_PRE;
 
     private int post = DEFAULT_POST;
-
-    static final SpanNotQueryBuilder PROTOTYPE = new SpanNotQueryBuilder(SpanTermQueryBuilder.PROTOTYPE, SpanTermQueryBuilder.PROTOTYPE);
 
     /**
      * Construct a span query matching spans from <code>include</code> which
@@ -125,14 +135,90 @@ public class SpanNotQueryBuilder extends AbstractQueryBuilder<SpanNotQueryBuilde
     @Override
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
-        builder.field(SpanNotQueryParser.INCLUDE_FIELD.getPreferredName());
+        builder.field(INCLUDE_FIELD.getPreferredName());
         include.toXContent(builder, params);
-        builder.field(SpanNotQueryParser.EXCLUDE_FIELD.getPreferredName());
+        builder.field(EXCLUDE_FIELD.getPreferredName());
         exclude.toXContent(builder, params);
-        builder.field(SpanNotQueryParser.PRE_FIELD.getPreferredName(), pre);
-        builder.field(SpanNotQueryParser.POST_FIELD.getPreferredName(), post);
+        builder.field(PRE_FIELD.getPreferredName(), pre);
+        builder.field(POST_FIELD.getPreferredName(), post);
         printBoostAndQueryName(builder);
         builder.endObject();
+    }
+
+    public static SpanNotQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+
+        SpanQueryBuilder include = null;
+        SpanQueryBuilder exclude = null;
+
+        Integer dist = null;
+        Integer pre  = null;
+        Integer post = null;
+
+        String queryName = null;
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, INCLUDE_FIELD)) {
+                    QueryBuilder query = parseContext.parseInnerQueryBuilder();
+                    if (!(query instanceof SpanQueryBuilder)) {
+                        throw new ParsingException(parser.getTokenLocation(), "spanNot [include] must be of type span query");
+                    }
+                    include = (SpanQueryBuilder) query;
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, EXCLUDE_FIELD)) {
+                    QueryBuilder query = parseContext.parseInnerQueryBuilder();
+                    if (!(query instanceof SpanQueryBuilder)) {
+                        throw new ParsingException(parser.getTokenLocation(), "spanNot [exclude] must be of type span query");
+                    }
+                    exclude = (SpanQueryBuilder) query;
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[span_not] query does not support [" + currentFieldName + "]");
+                }
+            } else {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, DIST_FIELD)) {
+                    dist = parser.intValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, PRE_FIELD)) {
+                    pre = parser.intValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, POST_FIELD)) {
+                    post = parser.intValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                    boost = parser.floatValue();
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                    queryName = parser.text();
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(), "[span_not] query does not support [" + currentFieldName + "]");
+                }
+            }
+        }
+        if (include == null) {
+            throw new ParsingException(parser.getTokenLocation(), "spanNot must have [include] span query clause");
+        }
+        if (exclude == null) {
+            throw new ParsingException(parser.getTokenLocation(), "spanNot must have [exclude] span query clause");
+        }
+        if (dist != null && (pre != null || post != null)) {
+            throw new ParsingException(parser.getTokenLocation(), "spanNot can either use [dist] or [pre] & [post] (or none)");
+        }
+
+        SpanNotQueryBuilder spanNotQuery = new SpanNotQueryBuilder(include, exclude);
+        if (dist != null) {
+            spanNotQuery.dist(dist);
+        }
+        if (pre != null) {
+            spanNotQuery.pre(pre);
+        }
+        if (post != null) {
+            spanNotQuery.post(post);
+        }
+        spanNotQuery.boost(boost);
+        spanNotQuery.queryName(queryName);
+        return spanNotQuery;
     }
 
     @Override
