+++ core/src/main/java/org/elasticsearch/index/query/SpanWithinQueryBuilder.java
@@ -22,9 +22,12 @@ package org.elasticsearch.index.query;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.spans.SpanQuery;
 import org.apache.lucene.search.spans.SpanWithinQuery;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -36,10 +39,15 @@ public class SpanWithinQueryBuilder extends AbstractQueryBuilder<SpanWithinQuery
         implements SpanQueryBuilder<SpanWithinQueryBuilder> {
 
     public static final String NAME = "span_within";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final SpanWithinQueryBuilder PROTOTYPE =
+            new SpanWithinQueryBuilder(SpanTermQueryBuilder.PROTOTYPE, SpanTermQueryBuilder.PROTOTYPE);
+
+    private static final ParseField BIG_FIELD = new ParseField("big");
+    private static final ParseField LITTLE_FIELD = new ParseField("little");
+
     private final SpanQueryBuilder big;
     private final SpanQueryBuilder little;
-    static final SpanWithinQueryBuilder PROTOTYPE =
-            new SpanWithinQueryBuilder(SpanTermQueryBuilder.PROTOTYPE, SpanTermQueryBuilder.PROTOTYPE);
 
     /**
      * Query that returns spans from <code>little</code> that are contained in a spans from <code>big</code>.
@@ -75,15 +83,66 @@ public class SpanWithinQueryBuilder extends AbstractQueryBuilder<SpanWithinQuery
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
 
-        builder.field(SpanWithinQueryParser.BIG_FIELD.getPreferredName());
+        builder.field(BIG_FIELD.getPreferredName());
         big.toXContent(builder, params);
 
-        builder.field(SpanWithinQueryParser.LITTLE_FIELD.getPreferredName());
+        builder.field(LITTLE_FIELD.getPreferredName());
         little.toXContent(builder, params);
 
         printBoostAndQueryName(builder);
 
         builder.endObject();
+    }
+
+    public static SpanWithinQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+        String queryName = null;
+        SpanQueryBuilder big = null;
+        SpanQueryBuilder little = null;
+
+        String currentFieldName = null;
+        XContentParser.Token token;
+        while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+            if (token == XContentParser.Token.FIELD_NAME) {
+                currentFieldName = parser.currentName();
+            } else if (token == XContentParser.Token.START_OBJECT) {
+                if (parseContext.parseFieldMatcher().match(currentFieldName, BIG_FIELD)) {
+                    QueryBuilder query = parseContext.parseInnerQueryBuilder();
+                    if (query instanceof SpanQueryBuilder == false) {
+                        throw new ParsingException(parser.getTokenLocation(), "span_within [big] must be of type span query");
+                    }
+                    big = (SpanQueryBuilder) query;
+                } else if (parseContext.parseFieldMatcher().match(currentFieldName, LITTLE_FIELD)) {
+                    QueryBuilder query = parseContext.parseInnerQueryBuilder();
+                    if (query instanceof SpanQueryBuilder == false) {
+                        throw new ParsingException(parser.getTokenLocation(), "span_within [little] must be of type span query");
+                    }
+                    little = (SpanQueryBuilder) query;
+                } else {
+                    throw new ParsingException(parser.getTokenLocation(),
+                            "[span_within] query does not support [" + currentFieldName + "]");
+                }
+            } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                boost = parser.floatValue();
+            } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                queryName = parser.text();
+            } else {
+                throw new ParsingException(parser.getTokenLocation(), "[span_within] query does not support [" + currentFieldName + "]");
+            }
+        }
+
+        if (big == null) {
+            throw new ParsingException(parser.getTokenLocation(), "span_within must include [big]");
+        }
+        if (little == null) {
+            throw new ParsingException(parser.getTokenLocation(), "span_within must include [little]");
+        }
+
+        SpanWithinQueryBuilder query = new SpanWithinQueryBuilder(big, little);
+        query.boost(boost).queryName(queryName);
+        return query;
     }
 
     @Override
