+++ core/src/main/java/org/elasticsearch/index/query/WildcardQueryBuilder.java
@@ -24,10 +24,13 @@ import org.apache.lucene.search.MultiTermQuery;
 import org.apache.lucene.search.Query;
 import org.apache.lucene.search.WildcardQuery;
 import org.apache.lucene.util.BytesRef;
+import org.elasticsearch.common.ParseField;
+import org.elasticsearch.common.ParsingException;
 import org.elasticsearch.common.Strings;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.xcontent.XContentBuilder;
+import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.index.mapper.MappedFieldType;
 import org.elasticsearch.index.query.support.QueryParsers;
 
@@ -46,14 +49,18 @@ public class WildcardQueryBuilder extends AbstractQueryBuilder<WildcardQueryBuil
         implements MultiTermQueryBuilder<WildcardQueryBuilder> {
 
     public static final String NAME = "wildcard";
+    public static final ParseField QUERY_NAME_FIELD = new ParseField(NAME);
+    public static final WildcardQueryBuilder PROTOTYPE = new WildcardQueryBuilder("field", "value");
+
+    private static final ParseField WILDCARD_FIELD = new ParseField("wildcard");
+    private static final ParseField VALUE_FIELD = new ParseField("value");
+    private static final ParseField REWRITE_FIELD = new ParseField("rewrite");
 
     private final String fieldName;
 
     private final String value;
 
     private String rewrite;
-
-    static final WildcardQueryBuilder PROTOTYPE = new WildcardQueryBuilder("field", "value");
 
     /**
      * Implements the wildcard search query. Supported wildcards are <tt>*</tt>, which
@@ -103,13 +110,64 @@ public class WildcardQueryBuilder extends AbstractQueryBuilder<WildcardQueryBuil
     protected void doXContent(XContentBuilder builder, Params params) throws IOException {
         builder.startObject(NAME);
         builder.startObject(fieldName);
-        builder.field(WildcardQueryParser.WILDCARD_FIELD.getPreferredName(), value);
+        builder.field(WILDCARD_FIELD.getPreferredName(), value);
         if (rewrite != null) {
-            builder.field(WildcardQueryParser.REWRITE_FIELD.getPreferredName(), rewrite);
+            builder.field(REWRITE_FIELD.getPreferredName(), rewrite);
         }
         printBoostAndQueryName(builder);
         builder.endObject();
         builder.endObject();
+    }
+
+    public static WildcardQueryBuilder fromXContent(QueryParseContext parseContext) throws IOException {
+        XContentParser parser = parseContext.parser();
+
+        XContentParser.Token token = parser.nextToken();
+        if (token != XContentParser.Token.FIELD_NAME) {
+            throw new ParsingException(parser.getTokenLocation(), "[wildcard] query malformed, no field");
+        }
+        String fieldName = parser.currentName();
+        String rewrite = null;
+
+        String value = null;
+        float boost = AbstractQueryBuilder.DEFAULT_BOOST;
+        String queryName = null;
+        token = parser.nextToken();
+        if (token == XContentParser.Token.START_OBJECT) {
+            String currentFieldName = null;
+            while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+                if (token == XContentParser.Token.FIELD_NAME) {
+                    currentFieldName = parser.currentName();
+                } else {
+                    if (parseContext.parseFieldMatcher().match(currentFieldName, WILDCARD_FIELD)) {
+                        value = parser.text();
+                    } else if (parseContext.parseFieldMatcher().match(currentFieldName, VALUE_FIELD)) {
+                        value = parser.text();
+                    } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.BOOST_FIELD)) {
+                        boost = parser.floatValue();
+                    } else if (parseContext.parseFieldMatcher().match(currentFieldName, REWRITE_FIELD)) {
+                        rewrite = parser.textOrNull();
+                    } else if (parseContext.parseFieldMatcher().match(currentFieldName, AbstractQueryBuilder.NAME_FIELD)) {
+                        queryName = parser.text();
+                    } else {
+                        throw new ParsingException(parser.getTokenLocation(),
+                                "[wildcard] query does not support [" + currentFieldName + "]");
+                    }
+                }
+            }
+            parser.nextToken();
+        } else {
+            value = parser.text();
+            parser.nextToken();
+        }
+
+        if (value == null) {
+            throw new ParsingException(parser.getTokenLocation(), "No value specified for prefix query");
+        }
+        return new WildcardQueryBuilder(fieldName, value)
+                .rewrite(rewrite)
+                .boost(boost)
+                .queryName(queryName);
     }
 
     @Override
